Reorganize by putting numeric and linalg code into submodules
