+++ src/linalg/impl_linalg.rs
@@ -40,54 +40,6 @@ const GEMM_BLAS_CUTOFF: usize = 7;
 #[allow(non_camel_case_types)]
 type blas_index = c_int; // blas index type
 
-impl<A, S> ArrayBase<S, (Ix, Ix)>
-    where S: Data<Elem=A>,
-{
-    /// Return an array view of row `index`.
-    ///
-    /// **Panics** if `index` is out of bounds.
-    pub fn row(&self, index: Ix) -> ArrayView<A, Ix>
-    {
-        self.subview(Axis(0), index)
-    }
-
-    /// Return a mutable array view of row `index`.
-    ///
-    /// **Panics** if `index` is out of bounds.
-    pub fn row_mut(&mut self, index: Ix) -> ArrayViewMut<A, Ix>
-        where S: DataMut
-    {
-        self.subview_mut(Axis(0), index)
-    }
-
-    /// Return the number of rows (length of `Axis(0)`) in the two-dimensional array.
-    pub fn rows(&self) -> Ix {
-        self.shape().axis(Axis(0))
-    }
-
-    /// Return an array view of column `index`.
-    ///
-    /// **Panics** if `index` is out of bounds.
-    pub fn column(&self, index: Ix) -> ArrayView<A, Ix>
-    {
-        self.subview(Axis(1), index)
-    }
-
-    /// Return a mutable array view of column `index`.
-    ///
-    /// **Panics** if `index` is out of bounds.
-    pub fn column_mut(&mut self, index: Ix) -> ArrayViewMut<A, Ix>
-        where S: DataMut
-    {
-        self.subview_mut(Axis(1), index)
-    }
-
-    /// Return the number of columns (length of `Axis(1)`) in the two-dimensional array.
-    pub fn cols(&self) -> Ix {
-        self.shape().axis(Axis(1))
-    }
-}
-
 
 impl<A, S> ArrayBase<S, Ix>
     where S: Data<Elem=A>,
